get rid of dependencycontrol bullshit
there is no sense in me keeping it around when all it does is
add faff to writing the script
add useless noise to the filenames/paths
and enforces its opinions upon me which i dont agree with

and all for absolutely ZERO reason since i don't even have a feed in the
first place

if you want to take my scripts and package them into a dependencycontrol
feed, you are more than welcome to do so, provided you comply with the
terms of the licence. but your life is going to be a little bit harder
i'm afraid, sorry about that.

i will keep all the script_namespace, depctrl registration and such 
around for the time being (removing it is >effort)